+++ CUDA/xfibres_gpu.cu
@@ -73,7 +73,7 @@ void xfibres_gpu(	//INPUT
 
 	if(nvox>0){
 		thrust::host_vector<float> datam_host, bvecs_host, bvals_host, params_host;
-		thrust::host_vector<float> alpha_host, beta_host;
+		thrust::host_vector<double> alpha_host, beta_host;
 		thrust::host_vector<float> tau_host;
 		vector<ColumnVector> datam_vec;
 		vector<Matrix> bvecs_vec, bvals_vec;
@@ -108,20 +108,20 @@ void xfibres_gpu(	//INPUT
 		int nVOX_multiple = int(nvox/VOXELS_BLOCK_MCMC)*VOXELS_BLOCK_MCMC;
 		if(nvox%VOXELS_BLOCK_MCMC) nVOX_multiple=nVOX_multiple+VOXELS_BLOCK_MCMC;
 
-		thrust::host_vector<float> signals_host;
-		thrust::host_vector<float> isosignals_host;
+		thrust::host_vector<double> signals_host;
+		thrust::host_vector<double> isosignals_host;
 		thrust::host_vector<FibreGPU> fibres_host;
 		thrust::host_vector<MultifibreGPU> multifibres_host;
 	
 		prepare_data_gpu_MCMC(nVOX_multiple, ndirections, nfib, signals_host, isosignals_host, fibres_host, multifibres_host);
 		
-		thrust::device_vector<float> signals_gpu=signals_host;
-		thrust::device_vector<float> isosignals_gpu=isosignals_host;
+		thrust::device_vector<double> signals_gpu=signals_host;
+		thrust::device_vector<double> isosignals_gpu=isosignals_host;
 		thrust::device_vector<FibreGPU> fibres_gpu=fibres_host;
 		thrust::device_vector<MultifibreGPU> multifibres_gpu=multifibres_host;
 		thrust::device_vector<float> tau_gpu = tau_host;
-		thrust::device_vector<float> alpha_gpu=alpha_host;
-		thrust::device_vector<float> beta_gpu=beta_host;
+		thrust::device_vector<double> alpha_gpu=alpha_host;
+		thrust::device_vector<double> beta_gpu=beta_host;
 
 		thrust::device_vector<curandState> randStates_gpu;
 		resize_structures(nVOX_multiple,ndirections,datam_gpu, params_gpu, tau_gpu, bvals_gpu, alpha_gpu, beta_gpu, randStates_gpu);
@@ -379,8 +379,8 @@ void prepare_data_gpu_FIT(	//INPUT
 				thrust::host_vector<float>&   		datam_host,	//data prepared for copy to GPU
 				thrust::host_vector<float>&		bvecs_host,				
 				thrust::host_vector<float>&		bvals_host,
-				thrust::host_vector<float>&		alpha_host,
-				thrust::host_vector<float>&		beta_host,
+				thrust::host_vector<double>&		alpha_host,
+				thrust::host_vector<double>&		beta_host,
 				thrust::host_vector<float>&		params_host,
 				thrust::host_vector<float>&		tau_host)
 {
@@ -580,8 +580,8 @@ void prepare_data_gpu_MCMC(	//INPUT
 				int					ndirections,
 				int 					nfib,
 				//OUTPUT
-				thrust::host_vector<float>&		signals_host,
-				thrust::host_vector<float>&		isosignals_host,
+				thrust::host_vector<double>&		signals_host,
+				thrust::host_vector<double>&		isosignals_host,
 				thrust::host_vector<FibreGPU>& 		fibres_host,
 				thrust::host_vector<MultifibreGPU>& 	multifibres_host)
 { 	
@@ -628,8 +628,8 @@ void resize_structures(		//INPUT
 				thrust::device_vector<float>&		params_gpu,
 				thrust::device_vector<float>&		tau_gpu,
 				thrust::device_vector<float>&		bvals_gpu,				
-				thrust::device_vector<float>&		alpha_gpu,
-				thrust::device_vector<float>&		beta_gpu,
+				thrust::device_vector<double>&		alpha_gpu,
+				thrust::device_vector<double>&		beta_gpu,
 				thrust::device_vector<curandState>&	randStates_gpu)
 {
 	xfibresOptions& opts = xfibresOptions::getInstance();
